reorganize operations
